requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
base58==0.2.5
bitarray==2.9.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.1
cytoolz==0.12.3
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.7.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
ethtoken==0.0.1a4
frozenlist==1.4.1
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
hexbytes==0.3.1
httplib2==0.22.0
idna==3.8
lru-dict==1.3.0
multidict==6.1.0
parsimonious==0.8.1
proto-plus==1.24.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
python-axolotl-curve25519==0.4.1.post2
python-dotenv==1.0.1
pyunormalize==15.1.0
PyWaves==1.0.5
regex==2024.9.11
requests==2.32.3
rlp==2.0.1
rsa==4.9
six==1.16.0
toolz==0.12.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
web3==4.10.0
websockets==6.0
yarl==1.11.1
